requirements.txt
torch
colpali-engine
transformers
langchain
qdrant-client
pillow
python-dotenv